backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pytest
httpx
web3
openai
python-dotenv
